ipc: update load_msg to take size_t
load_msg() originally took a signed int which could have resulted in a
kmalloc requesting less that the necessary size for msg and subsequently
overwriting pieces of kernel memory if the parameter was sufficiently
large to be considered a negative number for sizeof(*msg)+len.

+++ ipc/msgutil.c
@@ -44,12 +44,12 @@ struct msg_msgseg {
 #define DATALEN_MSG	(PAGE_SIZE-sizeof(struct msg_msg))
 #define DATALEN_SEG	(PAGE_SIZE-sizeof(struct msg_msgseg))
 
-struct msg_msg *load_msg(const void __user *src, int len)
+struct msg_msg *load_msg(const void __user *src, size_t len)
 {
 	struct msg_msg *msg;
 	struct msg_msgseg **pseg;
 	int err;
-	int alen;
+	size_t alen;
 
 	alen = len;
 	if (alen > DATALEN_MSG)
